% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Album_Functions.r
\name{get_album_summary}
\alias{get_album_summary}
\title{Search for album's feature summary}
\usage{
get_album_summary(id, authorization = get_spotify_access_token())
}
\arguments{
\item{id}{\itemize{
\item A single Spotify album id
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of album data, including album name and id, album popularity, number of tracks, and the mean and standard deviation for:
danceability
energy
loudness
speechiness
acousticness
instrumentalness
liveness
valence
tempo
duration_ms
mode
}
\description{
Search for album's feature summary
}
\examples{
\dontrun{
get_album_summary("1uyf3l2d4XYwiEqAb7t7fX")
}
}
