% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Album_Functions.r, R/Tracks_Functions.r
\name{get_album_track_features}
\alias{get_album_track_features}
\title{Search for Spotify album track features}
\usage{
get_album_track_features(id, authorization = get_spotify_access_token())

get_album_track_features(id, authorization = get_spotify_access_token())
}
\arguments{
\item{id}{\itemize{
\item A single Spotify album id
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of album data, including album name and id, album popularity, number of tracks, and the mean and standard deviation for:
danceability
energy
loudness
mode
speechiness
acousticness
instrumentalness
liveness
valence
tempo
duration_ms
time_signature

A dataframe of album track data, including ids and names for the tracks, whether the track is explicit, track number in the album, artist names and ids, and the following variables:
danceability
energy
key
loudness
mode
speechiness
acousticness
instrumentalness
liveness
valence
tempo
duration_ms
time_signature
}
\description{
Search for Spotify album track features

Search for Spotify album's track features
}
\examples{
\dontrun{
get_album_track_features("1uyf3l2d4XYwiEqAb7t7fX")
}
\dontrun{
get_album_track_features("58ufpQsJ1DS5kq4hhzQDiI")
}
}
