% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Album_Functions.r
\name{get_artist_projects}
\alias{get_artist_projects}
\title{Search for Spotify artist's projects}
\usage{
get_artist_projects(
  id,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{id}{\itemize{
\item A single Spotify artist id
}}

\item{limit}{\itemize{
\item Optional.  Number of albums wanted to return.  Valid if between 1 and 50.  Defaults to 20
}}

\item{offset}{\itemize{
\item Optional.  Index of first album wanted.  Defaults to 0
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of project data, including the project type, project id and name, release date, number of tracks, and the artists names and ids
}
\description{
Search for Spotify artist's projects
}
\examples{
\dontrun{
get_artist_projects("0du5cEVh5yTK9QJze8zA0C")
get_artist_projects("0du5cEVh5yTK9QJze8zA0C", limit = 50, offset = 2)
}
}
