% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artist_Functions.r
\name{get_artist_summary}
\alias{get_artist_summary}
\title{Search for Spotify artist feature summary}
\usage{
get_artist_summary(id, authorization = get_spotify_access_token())
}
\arguments{
\item{id}{\itemize{
\item A single Spotify artist id
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of artist summary data, including artist name and id, number of songs from the artist, and the mean and standard deviation for:
danceability
energy
loudness
speechiness
acousticness
instrumentalness
liveness
valence
tempo
duration_ms
mode
}
\description{
Search for Spotify artist feature summary
}
\examples{
\dontrun{
get_artist_summary("0du5cEVh5yTK9QJze8zA0C")
}
}
