% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artist_Functions.r
\name{get_artists}
\alias{get_artists}
\title{Search for Spotify artist information}
\usage{
get_artists(ids, authorization = get_spotify_access_token())
}
\arguments{
\item{ids}{\itemize{
\item A vector of Spotify artist ids
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of artist data, including artist name and id, genres of the artist, artist popularity, and the total number of followers on spotify
}
\description{
Search for Spotify artist information
}
\examples{
\dontrun{
get_artists("0du5cEVh5yTK9QJze8zA0C")
get_artists(c("0du5cEVh5yTK9QJze8zA0C", "6PvvGcCY2XtUcSRld1Wilr"))
}
}
