% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artist_Functions.r
\name{get_artists_summary}
\alias{get_artists_summary}
\title{Search for Spotify artists feature summary}
\usage{
get_artists_summary(ids, authorization = get_spotify_access_token())
}
\arguments{
\item{ids}{\itemize{
\item A vector of Spotify artist ids
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of artist summary data, including artist name and id, number of songs from the artist, and the mean and standard deviation for:
danceability
energy
loudness
speechiness
acousticness
instrumentalness
liveness
valence
tempo
duration_ms
mode
}
\description{
Search for Spotify artists feature summary
}
\examples{
\dontrun{
get_artists_summary(c("0du5cEVh5yTK9QJze8zA0C", "6PvvGcCY2XtUcSRld1Wilr"))
}
}
