% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tracks_Functions.r
\name{get_track_bars}
\alias{get_track_bars}
\title{Search for Spotify track bar information}
\usage{
get_track_bars(id, authorization = get_spotify_access_token())
}
\arguments{
\item{id}{\itemize{
\item A single Spotify track id
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of track bar information, including track id, and start, end, and confidence of bar intervals
}
\description{
Search for Spotify track bar information
}
\examples{
\dontrun{
get_track_bars("3w3y8KPTfNeOKPiqUTakBh")
}
}
