% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tracks_Functions.r
\name{get_track_beats}
\alias{get_track_beats}
\title{Search for Spotify track beats information}
\usage{
get_track_beats(id, authorization = get_spotify_access_token())
}
\arguments{
\item{id}{\itemize{
\item A single Spotify track id
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of track beat information, including track id, and start, end, and confidence of beat intervals
}
\description{
Search for Spotify track beats information
}
\examples{
\dontrun{
get_track_beats("3w3y8KPTfNeOKPiqUTakBh")
}
}
