% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tracks_Functions.r
\name{get_track_sections}
\alias{get_track_sections}
\title{Search for Spotify track section information}
\usage{
get_track_sections(id, authorization = get_spotify_access_token())
}
\arguments{
\item{id}{\itemize{
\item A single Spotify track id
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of track section information, including track id, start, end, and confidence of section intervals, loudness, tempo and confidence, key and confidence, mode and confidence, and time signature and confidence
}
\description{
Search for Spotify track section information
}
\examples{
\dontrun{
get_track_sections("3w3y8KPTfNeOKPiqUTakBh")
}
}
